Added Services Package with a lot of classes. Fully functional rental system Fully functional search system Fully functional rent availability Added Statistics for a specific CAR (only that ftm) Added Confirm Pending Transaction Added Confirm Car Rental Return Null-proofed inputs

+++ src/ro/jademy/carrental/models/cars/bmw/Bmw.java
@@ -3,7 +3,7 @@ import ro.jademy.carrental.models.cars.Car;
 
 public class Bmw extends Car {
 
-    public Bmw(int id, String model, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, int basePrice) {
+    public Bmw(int id, String model, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, long basePrice) {
         super(id, "Bmw", model, variant, carType, year, engine, fuelType, transmissionType, doorNumber, color, VIN, basePrice);
     }
 }
