+++ src/ro/jademy/carrental/models/cars/bmw/Serie5.java
@@ -5,7 +5,7 @@ package ro.jademy.carrental.models.cars.bmw;
 public class Serie5 extends Bmw{
 
 
-    public Serie5(int id,  String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, int basePrice) {
+    public Serie5(int id, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, long basePrice) {
         super(id, "Serie5", variant, carType, year, engine, fuelType, transmissionType, doorNumber, color, VIN, basePrice);
     }
 }
