+++ src/ro/jademy/carrental/models/cars/bmw/Serie7.java
@@ -5,7 +5,7 @@ package ro.jademy.carrental.models.cars.bmw;
 public class Serie7 extends Bmw {
 
 
-    public Serie7(int id,  String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, int basePrice) {
+    public Serie7(int id, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, long basePrice) {
         super(id, "Serie7", variant, carType, year, engine, fuelType, transmissionType, doorNumber, color, VIN, basePrice);
     }
 }
