+++ src/ro/jademy/carrental/models/cars/dacia/Dacia.java
@@ -4,7 +4,7 @@ import ro.jademy.carrental.models.cars.Car;
 public class Dacia extends Car {
 
 
-    public Dacia(int id,  String model, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, int basePrice) {
+    public Dacia(int id, String model, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, long basePrice) {
         super(id, "Dacia", model, variant, carType, year, engine, fuelType, transmissionType, doorNumber, color, VIN, basePrice);
     }
 }
