+++ src/ro/jademy/carrental/models/cars/dacia/Duster.java
@@ -6,7 +6,7 @@ package ro.jademy.carrental.models.cars.dacia;
 public class Duster extends Dacia {
 
 
-    public Duster(int id,  String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, int basePrice) {
+    public Duster(int id, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, long basePrice) {
         super(id, "Duster", variant, carType, year, engine, fuelType, transmissionType, doorNumber, color, VIN, basePrice);
     }
 }
