+++ src/ro/jademy/carrental/models/cars/dacia/Logan.java
@@ -4,7 +4,7 @@ package ro.jademy.carrental.models.cars.dacia;
 public class Logan extends Dacia{
 
 
-    public Logan(int id,  String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, int basePrice) {
+    public Logan(int id, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, long basePrice) {
         super(id, "Logan", variant, carType, year, engine, fuelType, transmissionType, doorNumber, color, VIN, basePrice);
     }
 }
