+++ src/ro/jademy/carrental/models/cars/dacia/Sandero.java
@@ -4,7 +4,7 @@ package ro.jademy.carrental.models.cars.dacia;
 public class Sandero extends Dacia {
 
 
-    public Sandero(int id, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, int basePrice) {
+    public Sandero(int id, String variant, String carType, int year, int engine, String fuelType, String transmissionType, Integer doorNumber, String color, String VIN, long basePrice) {
         super(id, "Sandero", variant, carType, year, engine, fuelType, transmissionType, doorNumber, color, VIN, basePrice);
     }
 }
